Attempted a few game play elements.
I have attempted the treasure deck.

Today's session, I have figured that we need to have a clear logic route on how to move forward. The setup of our game is done however we have need a clear logic route moving forward. I attempted a few gameplay elements and have commented out the code in case we need it for later use.